Modify customer entity reg. membership
