+++ src/main/java/com/rkasibha/rentabook/dto/CustomerBasicInfoDto.java
@@ -1,6 +1,9 @@
 package com.rkasibha.rentabook.dto;
 
+import com.rkasibha.rentabook.enumeration.MembershipStatus;
 import lombok.Data;
+
+import java.time.LocalDate;
 
 @Data
 public class CustomerBasicInfoDto {
@@ -9,4 +12,6 @@ public class CustomerBasicInfoDto {
     private String lastName;
     private String contactNumber;
     private String emailId;
+    private LocalDate memberSince;
+    private MembershipStatus membershipStatus;
 }
