+++ src/main/java/com/rkasibha/rentabook/exception/DataNotFoundException.java
@@ -1,8 +1,10 @@
 package com.rkasibha.rentabook.exception;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @Data
+@EqualsAndHashCode(callSuper = false)
 public class DataNotFoundException extends Exception {
     private String message;
 }
