--- src/main/java/com/rkasibha/rentabook/model/Membership.java
@@ -1,70 +1,0 @@
-package com.rkasibha.rentabook.model;
-
-import com.rkasibha.rentabook.enumeration.MembershipStatus;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
-import java.time.LocalDate;
-
-@Entity
-public class Membership {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @NotNull
-    private LocalDate memberSince;
-
-    @Enumerated(EnumType.STRING)
-    private MembershipStatus status;
-
-    @ManyToOne
-    private Branch branch;
-
-    @ManyToOne
-    private Customer customer;
-
-    public Membership() {
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public LocalDate getMemberSince() {
-        return memberSince;
-    }
-
-    public void setMemberSince(LocalDate memberSince) {
-        this.memberSince = memberSince;
-    }
-
-    public MembershipStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(MembershipStatus status) {
-        this.status = status;
-    }
-
-    public Branch getBranch() {
-        return branch;
-    }
-
-    public void setBranch(Branch branch) {
-        this.branch = branch;
-    }
-
-    public Customer getCustomer() {
-        return customer;
-    }
-
-    public void setCustomer(Customer customer) {
-        this.customer = customer;
-    }
-}
