--- src/main/java/com/rkasibha/rentabook/repository/MembershipRepository.java
@@ -1,9 +1,0 @@
-package com.rkasibha.rentabook.repository;
-
-import com.rkasibha.rentabook.model.Membership;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MembershipRepository extends JpaRepository<Membership, Integer> {
-}
